+++ rocketgram/api/update.py
@@ -3,6 +3,7 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from typing import Dict, Optional
 
@@ -26,12 +27,12 @@ class Update:
 
     Additional fields:
     raw
-    update_type
+    type
     """
 
     raw: Dict
     update_id: int
-    update_type: UpdateType
+    type: UpdateType
     message: Optional[Message]
     edited_message: Optional[Message]
     channel_post: Optional[Message]
@@ -93,3 +94,9 @@ class Update:
         return cls(data, data['update_id'], update_type, message, edited_message, channel_post, edited_channel_post,
                    inline_query, chosen_inline_result, callback_query, shipping_query, pre_checkout_query, poll,
                    poll_answer, my_chat_member, chat_member)
+
+    @property
+    def update_type(self) -> UpdateType:
+        warnings.warn("This field is deprecated. Use `type` instead.", DeprecationWarning)
+
+        return self.type