+++ rocketgram/api/user.py
@@ -3,6 +3,7 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+import warnings
 from dataclasses import dataclass
 from typing import Dict, Optional
 
@@ -12,12 +13,9 @@ class User:
     """\
     Represents User object:
     https://core.telegram.org/bots/api#user
-
-    Differences in field names:
-    id -> user_id
     """
 
-    user_id: int
+    id: int
     is_bot: bool
     first_name: str
     last_name: Optional[str]
@@ -35,3 +33,9 @@ class User:
         return cls(data['id'], data['is_bot'], data['first_name'], data.get('last_name'), data.get('username'),
                    data.get('language_code'), data.get('can_join_groups'), data.get('can_read_all_group_messages'),
                    data.get('supports_inline_queries'))
+
+    @property
+    def user_id(self) -> int:
+        warnings.warn("This field is deprecated. Use `id` instead.", DeprecationWarning)
+
+        return self.id